# docs/public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Asmblyr Core HTTP API",
    "version": "0.0.0",
    "description": "Все статические маршруты Core и их требования доступа. Часть операций имеет только route-only описание; спецификация не является полной схемой для генерации клиента. OAuth protocol и динамические маршруты плагинов описаны отдельно."
  },
  "servers": [
    {
      "url": "http://localhost:3001",
      "description": "Локальный Core"
    }
  ],
  "paths": {
...
    "/item-events/{collection}": {
      "get": {
        "tags": [
          "item-events"
        ],
        "operationId": "get__item_events_collection",
        "summary": "GET /item-events/:collection",
        "description": "Активный principal; действие и поля коллекции, связанные данные и история проверяются отдельно. Схемы тела и ответа пока не детализированы; см. реализацию обработчика. Источник: apps/core/src/items/routes.ts",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "collection",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "description": "Контракт ответа смотрите в обработчике; эта операция описана на уровне маршрута и доступа."
          }
        },
        "x-contract-level": "route-only",
        "x-source": "apps/core/src/items/routes.ts"
      }
    },
    "/items/{collection}": {
      "get": {
        "tags": [
          "items"
        ],
        "operationId": "get__items_collection",
        "summary": "GET /items/:collection",
        "description": "Активный principal; действие и поля коллекции, связанные данные и история проверяются отдельно. Источник: apps/core/src/items/routes.ts",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "collection",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "fields",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "Имена через запятую; пустое значение возвращает только ключ. Пропуск — все доступные поля."
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 100
            }
          },
          {
            "name": "sort",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "direction",
            "in": "query",
            "schema": {
              "enum": [
                "asc",
                "desc"
              ]
            }
          },
          {
            "name": "q",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "Поиск по настроенным полям и связям."
          },
          {
            "name": "filter",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "JSON группы {logic: and|or, children: [...]}; это фильтр запроса, не grant."
          }
        ],
        "responses": {
          "200": {
            "description": "Успех",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "additionalProperties": true
                      }
                    },
                    "labels": {
                      "type": "object",
                      "additionalProperties": {
                        "type": "string"
                      }
                    },
                    "page": {
                      "type": "object",
                      "properties": {
                        "number": {
                          "type": "integer"
                        },
                        "size": {
                          "type": "integer"
                        },
                        "total": {
                          "type": "string"
                        },
                        "sort": {
                          "type": "string"
                        },
                        "direction": {
                          "enum": [
                            "asc",
                            "desc"
                          ]
                        }
                      },
                      "required": [
                        "number",
                        "size",
                        "total",
                        "sort",
                        "direction"
                      ]
                    }
                  },
                  "required": [
                    "data",
                    "labels",
                    "page"
                  ]
                }
              }
            }
          },
          "default": {
            "description": "Ошибка проверки, авторизации или исполнения; код HTTP определяет результат."
          }
        },
        "x-contract-level": "documented",
        "x-source": "apps/core/src/items/routes.ts"
      },
      "patch": {
        "tags": [
          "items"
        ],
        "operationId": "patch__items_collection",
        "summary": "PATCH /items/:collection",
        "description": "Активный principal; действие и поля коллекции, связанные данные и история проверяются отдельно. Схемы тела и ответа пока не детализированы; см. реализацию обработчика. Источник: apps/core/src/items/routes.ts",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "collection",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "description": "Контракт ответа смотрите в обработчике; эта операция описана на уровне маршрута и доступа."
          }
        },
        "x-contract-level": "route-only",
        "x-source": "apps/core/src/items/routes.ts"
      },
      "post": {
        "tags": [
          "items"
        ],
        "operationId": "post__items_collection",
        "summary": "POST /items/:collection",
        "description": "Активный principal; действие и поля коллекции, связанные данные и история проверяются отдельно. Источник: apps/core/src/items/routes.ts",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "collection",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Сохранено; data=null при отсутствии права чтения.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "anyOf": [
                        {
                          "type": "object",
                          "additionalProperties": true
                        },
                        {
                          "type": "null"
                        }
                      ]
                    }
                  },
                  "required": [
                    "data"
                  ]
                }
              }
            }
          },
          "default": {
            "description": "Ошибка проверки, авторизации или исполнения; код HTTP определяет результат."
          }
        },
        "x-contract-level": "documented",
        "x-source": "apps/core/src/items/routes.ts",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true
              }
            }
          }
        }
      }
    },
    "/items/{collection}/{id}": {
      "delete": {
        "tags": [
          "items"
        ],
        "operationId": "delete__items_collection_id",
        "summary": "DELETE /items/:collection/:id",
        "description": "Активный principal; действие и поля коллекции, связанные данные и история проверяются отдельно. Источник: apps/core/src/items/routes.ts",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "collection",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Удалено, без тела ответа."
          },
          "default": {
            "description": "Ошибка проверки, авторизации или исполнения; код HTTP определяет результат."
          }
        },
        "x-contract-level": "documented",
        "x-source": "apps/core/src/items/routes.ts"
      },
      "get": {
        "tags": [
          "items"
        ],
        "operationId": "get__items_collection_id",
        "summary": "GET /items/:collection/:id",
        "description": "Активный principal; действие и поля коллекции, связанные данные и история проверяются отдельно. Источник: apps/core/src/items/routes.ts",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "collection",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "fields",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "Имена через запятую; пустое значение возвращает только ключ. Пропуск — все доступные поля."
          }
        ],
        "responses": {
          "200": {
            "description": "Успех",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "additionalProperties": true
                    },
                    "label": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "data",
                    "label"
                  ]
                }
              }
            }
          },
          "default": {
            "description": "Ошибка проверки, авторизации или исполнения; код HTTP определяет результат."
          }
        },
        "x-contract-level": "documented",
        "x-source": "apps/core/src/items/routes.ts"
      },
      "patch": {
        "tags": [
          "items"
        ],
        "operationId": "patch__items_collection_id",
        "summary": "PATCH /items/:collection/:id",
        "description": "Активный principal; действие и поля коллекции, связанные данные и история проверяются отдельно. Источник: apps/core/src/items/routes.ts",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "collection",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Сохранено; data=null при отсутствии права чтения.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "anyOf": [
                        {
                          "type": "object",
                          "additionalProperties": true
                        },
                        {
                          "type": "null"
                        }
                      ]
                    }
                  },
                  "required": [
                    "data"
                  ]
                }
              }
            }
          },
          "default": {
            "description": "Ошибка проверки, авторизации или исполнения; код HTTP определяет результат."
          }
        },
        "x-contract-level": "documented",
        "x-source": "apps/core/src/items/routes.ts",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true
              }
            }
          }
        }
      }
    },
    "/items/{collection}/{id}/related": {
      "get": {
        "tags": [
          "items"
        ],
        "operationId": "get__items_collection_id_related",
        "summary": "GET /items/:collection/:id/related",
        "description": "Активный principal; действие и поля коллекции, связанные данные и история проверяются отдельно. Схемы тела и ответа пока не детализированы; см. реализацию обработчика. Источник: apps/core/src/items/routes.ts",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "collection",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "description": "Контракт ответа смотрите в обработчике; эта операция описана на уровне маршрута и доступа."
          }
        },
        "x-contract-level": "route-only",
        "x-source": "apps/core/src/items/routes.ts"
      }
    },
...
    "/search": {
      "get": {
        "tags": [
          "search"
        ],
        "operationId": "get__search",
        "summary": "GET /search",
        "description": "Активный principal; действие и поля коллекции, связанные данные и история проверяются отдельно. Схемы тела и ответа пока не детализированы; см. реализацию обработчика. Источник: apps/core/src/items/routes.ts",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [],
        "responses": {
          "default": {
            "description": "Контракт ответа смотрите в обработчике; эта операция описана на уровне маршрута и доступа."
          }
        },
        "x-contract-level": "route-only",
        "x-source": "apps/core/src/items/routes.ts"
      }
    },
...
  },
  "components": {
...
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Непрозрачный access token пользователя или сервиса. Разделы настроек доступны только человеку."
      }
    }
  },
...
}